requirements.txt
streamlit  
openai-whisper
transformers 
sentence-transformers 
torch 
faiss-cpu 
opencv-python
yt-dlp 
pillow
scikit-learn 
rank-bm25 
pgvector
psycopg2-binary 
sqlalchemy 
ffmpeg-python
tf-keras 
psycopg2 